added organization
moved files into different folders for organization purposes
