+++ rizz/lib/browseChat/browsing.dart
@@ -7,8 +7,8 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'consts.dart';
-import 'userObjects.dart';
+import '../consts.dart';
+import '../userObjects.dart';
 
 class LifestyleDisplayItem extends StatelessWidget {
   final String lifestyle;
