+++ rizz/lib/main.dart
@@ -4,7 +4,7 @@ library main;
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'firebase_options.dart';
-import 'login.dart';
+import 'login/login.dart';
 import 'style.dart';
 
 /// [main] intializes Firebase, then calls
