+++ rizz/lib/profile/user_profile.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'userObjects.dart';
-import 'name.dart';
-import 'login.dart';
+import '../userObjects.dart';
+import '../questionnaire/name.dart';
+import '../login/login.dart';
 
 const bkgSVG = '''
 <svg xmlns="http://www.w3.org/2000/svg" width="393" height="411" viewBox="0 0 393 411" fill="none">
