+++ rizz/lib/questionnaire/lifestyle.dart
@@ -5,10 +5,10 @@ library lifestyle.dart;
 
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:rizz/seeking.dart';
-import 'consts.dart';
+import 'seeking.dart';
+import '../consts.dart';
 import 'genderlifestyleinfo.dart';
-import 'nextbutton.dart';
+import '../nextbutton.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 
 /// Used to create a
