+++ rizz/lib/questionnaire/lifestylequiz.dart
@@ -5,9 +5,9 @@ library lifestylequiz.dart;
 
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'consts.dart';
+import '../consts.dart';
 import 'genderlifestyleinfo.dart';
-import 'nextbutton.dart';
+import '../nextbutton.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'aboutme.dart';
