+++ rizz/lib/questionnaire/photos.dart
@@ -3,9 +3,9 @@ library photos.dart;
 
 import 'package:flutter/material.dart';
 import 'package:rizz/header.dart';
-import 'spacingbox.dart';
-import 'consts.dart';
-import 'nextbutton.dart';
+import '../spacingbox.dart';
+import '../consts.dart';
+import '../nextbutton.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
