+++ rizz/lib/questionnaire/seeking.dart
@@ -1,52 +1,54 @@
 /// This page is about the genders that the user
 /// is seeking out for in others.
 library seeking.dart;
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:rizz/nextbutton.dart';
-import 'consts.dart';
+import '../consts.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'genderlifestyleinfo.dart';
 import 'lifestylequiz.dart';
 
 /// Used for making the Seeking Page.
-class SeekingPage extends StatefulWidget{
+class SeekingPage extends StatefulWidget {
   const SeekingPage({Key? key}) : super(key: key);
 
   @override
   State<StatefulWidget> createState() => SeekingPageState();
 }
 
-/// Makes the Seeking Page. 
+/// Makes the Seeking Page.
 /// Contains text, a list of checkboxes, and the confirm button.
-class SeekingPageState extends State<SeekingPage>{
-
+class SeekingPageState extends State<SeekingPage> {
   @override
   void initState() {
     super.initState();
   }
 
   List<GenderCheckbox> checkboxList = Gender.getGenders();
-  List<String> getGender(){
+  List<String> getGender() {
     List<String> genders = [];
-    for(int i = 0; i < checkboxList.length; i++){
-      if(checkboxList[i].value){
-         genders.add(checkboxList[i].gender);
+    for (int i = 0; i < checkboxList.length; i++) {
+      if (checkboxList[i].value) {
+        genders.add(checkboxList[i].gender);
       }
     }
     return genders;
   }
 
   // function to upload data to database
-  void uploadSeekingToDatabase() async{
+  void uploadSeekingToDatabase() async {
     User? user = FirebaseAuth.instance.currentUser;
     final db = FirebaseFirestore.instance;
-    await db.collection('users').doc(user!.uid).update({"seeking": getGender()});
+    await db
+        .collection('users')
+        .doc(user!.uid)
+        .update({"seeking": getGender()});
   }
 
-
   @override
-  Widget build(BuildContext context){
+  Widget build(BuildContext context) {
     return Scaffold(
       body: Container(
         color: Theme.of(context).colorScheme.background,
@@ -64,38 +66,34 @@ class SeekingPageState extends State<SeekingPage>{
               flex: 1,
               child: ListView.builder(
                 itemCount: checkboxList.length,
-                itemBuilder: (BuildContext context, int index){
+                itemBuilder: (BuildContext context, int index) {
                   final checkbox = checkboxList[index];
                   return CheckboxListTile(
-                    value: checkbox.value,
-                    title: Text(checkbox.gender), 
-                    onChanged: (value){
-                      setState(() {
-                        checkbox.value = value ?? false;
+                      value: checkbox.value,
+                      title: Text(checkbox.gender),
+                      onChanged: (value) {
+                        setState(() {
+                          checkbox.value = value ?? false;
+                        });
+                        print("changing value to $value");
                       });
-                      print("changing value to $value");
-                    }
-                  );
                 },
               ),
             ),
-            
             Container(
               margin: Consts.bottomButtonPadding,
-              child: NextButton(
-                onPressed: (){
-                  uploadSeekingToDatabase();
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => const LifestyleQuizPage()),
-                  );
-
-                }
-              ),
+              child: NextButton(onPressed: () {
+                uploadSeekingToDatabase();
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                      builder: (context) => const LifestyleQuizPage()),
+                );
+              }),
             ),
           ],
         ),
       ),
-    );    
+    );
   }
 }
